# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cancun Chaos Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /traits/{playerId}:
    get:
      operationId: getPlayerTraits
      tags: [players]
      summary: Get player's personality traits
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player traits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerTrait"
components:
  schemas:
...
    PlayerTrait:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        traitId:
          type: integer
        traitName:
          type: string
        traitEmoji:
          type: string
        level:
          type: integer
      required:
        - id
        - playerId
        - traitId
        - traitName
        - traitEmoji
        - level
...